requirements.txt
yt-dlp>=2025.7.21
groq>=0.31.0
python-dotenv>=1.0.1
jinja2>=3.1.2